% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccination.R
\name{apply_vaccination_patterns}
\alias{apply_vaccination_patterns}
\title{Apply Vaccination Patterns}
\usage{
apply_vaccination_patterns(centrality_measures, data, data.size, treatments)
}
\arguments{
\item{centrality_measures}{alist of igraph centrality measure}

\item{data}{An adjacency matrix, possibly directed}

\item{data.size}{Number of columns in the matrix}

\item{treatments}{Number of treatments available}
}
\value{
List of Lists with all the vaccination vectors [[1]] and
the vaccinated matrices [[2]] in the order of the centrality_measures
list.
}
\description{
A function that applies different vaccination patterns to a 
network.
}
\examples{
mat = matrix(rbinom(10 * 5, 1, 0.3), ncol = 10, nrow = 10)
vaccinatedIndex <- apply_vaccination_patterns(list(igraph::betweenness,igraph::closeness),mat,10,2)

}
\keyword{COmparison}
\keyword{Centrality,}
\keyword{Network,}

