% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection.R
\name{compare_infections}
\alias{compare_infections}
\title{Compare Infections}
\usage{
compare_infections(unvac.network, vac.network, intervention, infection.steps)
}
\arguments{
\item{unvac.network}{Network with no vaccinations}

\item{vac.network}{Network with vaccinations}

\item{intervention}{Vector of nodes representing vaccinated people}

\item{infection.steps}{Number of steps that the infection will spread.}
}
\value{
List with the graph representation of the unvaccinated network [[1]],
the vaccinated network [[2]], the infected index under no vaccination [[3]],
and the infected index under vaccination [[4]].
}
\description{
A function that takes takes two copies of a network,
one vaccinated and one unvaccinated
}
\examples{
n=100
mat <- rgraph(n, mode = "graph", tprob = 0.5)
net <- network(mat, directed = FALSE)
results <- identify_vaccination_targets(igraph::betweenness,mat,n,5)
intervention <- results[[1]]
vaccinatedMat <- do.call(rbind,results[2])
vacNet <- network(vaccinatedMat, directed = FALSE)
results2 <- compare_infections(net,vacNet,intervention,3)

}
\keyword{Comparison}
\keyword{Infection,}
\keyword{Network,}

