% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccination.R
\name{identify_vaccination_targets}
\alias{identify_vaccination_targets}
\title{Identify Vaccination Targets}
\usage{
identify_vaccination_targets(centrality_measure, data, network.size, treatments)
}
\arguments{
\item{centrality_measure}{an igraph centrality measure}

\item{data}{An adjacency matrix, possibly directed}
}
\value{
vector stating for each node whether it is "Vaccinated" or "Untreated".
}
\description{
A function that finds vaccination targets based on some network
centrality measure.
}
\examples{
mat = matrix(rbinom(10 * 5, 1, 0.3), ncol = 10, nrow = 10)
vaccinatedIndex <- identify_vaccination_targets(igraph::betweenness,mat)

}
\keyword{Centrality}
\keyword{Network,}

