% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection.R
\name{infect_network}
\alias{infect_network}
\title{Infect Network}
\usage{
infect_network(network, repetitions, network.size, random.seed = NULL)
}
\arguments{
\item{network}{An undirected network object}

\item{repetitions}{Number of steps the infection will spread}

\item{network.size}{Number of individuals in the network}

\item{random.seed}{A fixed seed for random infection.}
}
\value{
vector stating for each node whether it is "Healthy" or "Infected".
}
\description{
A function that infects a random individual of a network and
then spreads the infection stepwise through the network return-
ing a state vector of infection.
}
\examples{
n=100
net = = network(rgraph(n, mode = "graph", tprob = 0.5), directed = FALSE)
diseasedIndex <- infect_network(net,5,n)

}
\keyword{Infection,}
\keyword{Network,}
\keyword{Random}

