% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection.R
\name{test_random_infections}
\alias{test_random_infections}
\title{Test random Infections}
\usage{
test_random_infections(unvac.network, vac.network, intervention, network.size,
  infection.steps, random.samples)
}
\arguments{
\item{unvac.network}{Network with no vaccinations}

\item{vac.network}{Network with vaccinations}

\item{intervention}{Vector of nodes representing vaccinated people}

\item{network.size}{Number of nodes in the network.}

\item{infection.steps}{Number of steps that the infection will spread.}

\item{random.samples}{Vector of nodes that will be infected.}
}
\value{
List with the number of infection in the unvaccinated network [[1]]
and the vaccinated network [[2]].
}
\description{
A function that takes takes two copies of a network,
one vaccinated and one unvaccinated and test random
infection in it.
}
\details{
For each random sample the infection will be simulated
for a the specified amount of steps.
}
\examples{
n=100
mat <- rgraph(n, mode = "graph", tprob = 0.5)
net <- network(mat, directed = FALSE)
results <- identify_vaccination_targets(igraph::betweenness,mat,n,5)
intervention <- results[[1]]
vaccinatedMat <- do.call(rbind,results[2])
vacNet <- network(vaccinatedMat, directed = FALSE)
seed <- sample.int(n,10,replace = TRUE)
results2 <- test_random_infections(net,vacNet,intervention,n,3,seed)

}
\keyword{Comparison}
\keyword{Infection,}
\keyword{Network,}

